fixed kya image loading
