+++ mobile/pubspec.yaml
@@ -14,7 +14,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # android test
-version: 2.0.1+20004
+version: 2.0.1+20005
 
 environment:
   sdk: '>=2.12.0 <3.0.0'
